\name{kernelPSI-package}
\alias{kernelPSI-package}
\alias{kernelPSI}
\docType{package}
\title{
\packageTitle{kernelPSI}
}
\description{
\packageDescription{kernelPSI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kernelPSI}
\packageIndices{kernelPSI}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{kernelPSI}

Maintainer: \packageMaintainer{kernelPSI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
